% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytoarray-layout.R
\name{ary028_layout}
\alias{ary028_layout}
\title{Built-in layout for the R&D Systems Proteome Profiler Mouse XL array (ARY028)}
\usage{
ary028_layout()
}
\value{
a list with `ref`, `ref_pos`, `neg_pos`, `analyte_map`, `splits`
}
\description{
Built-in layout for the R&D Systems Proteome Profiler Mouse XL array (ARY028)
}
