% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytoarray-image.R
\name{compute_affine}
\alias{compute_affine}
\title{Fit the template-to-membrane affine map.}
\usage{
compute_affine(ref_template, ref_membrane)
}
\description{
Fit the template-to-membrane affine map.
}
