% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytoarray-image.R
\name{draw_overlay}
\alias{draw_overlay}
\title{Draw a verification overlay.}
\usage{
draw_overlay(
  img,
  spots,
  save_path = NULL,
  original_path = NULL,
  high_intensity_quantile = 0.75
)
}
\description{
Draw a verification overlay.
}
