% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytoarray-image.R
\name{map_spots}
\alias{map_spots}
\title{Map template spots onto the membrane.}
\usage{
map_spots(template_spots, M, uniform_radius = NULL, min_radius_px = 4)
}
\description{
Map template spots onto the membrane.
}
