% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytoarray-image.R
\name{preprocess_image}
\alias{preprocess_image}
\title{Read and background-flatten a membrane image}
\usage{
preprocess_image(img_path, brush = 51, invert_if_mean_above = 0.5)
}
\arguments{
\item{brush}{opening-disc diameter in pixels.}

\item{invert_if_mean_above}{invert scans brighter than this mean intensity.}
}
\description{
Read and background-flatten a membrane image
}
